use update parameter instead of setting a new one
