+++ app/static/src/app/components/options/OptionsTab.vue
@@ -4,12 +4,6 @@
         <vertical-collapse v-if="isFit" title="Link" collapse-id="link-data">
           <link-data></link-data>
         </vertical-collapse>
-        <div>
-            <button class="btn btn-primary btn-sm mb-2"
-                    id="reset-params-btn"
-                    @click="reset">Reset
-            </button>
-        </div>
         <vertical-collapse title="Model Parameters" collapse-id="model-params">
           <parameter-values></parameter-values>
         </vertical-collapse>
@@ -32,8 +26,6 @@ import LinkData from "./LinkData.vue";
 import SensitivityOptions from "./SensitivityOptions.vue";
 import OptimisationOptions from "./OptimisationOptions.vue";
 import { AppType, VisualisationTab } from "../../store/appState/state";
-import { RunMutation } from "../../store/run/mutations";
-import { OdinParameter, OdinUserType } from "../../types/responseTypes";
 
 export default {
     name: "OptionsTab",
@@ -51,20 +43,8 @@ export default {
 
         const sensitivityOpen = computed(() => store.state.openVisualisationTab === VisualisationTab.Sensitivity);
         const fitTabIsOpen = computed(() => store.state.openVisualisationTab === VisualisationTab.Fit);
-        const odinParameters = computed(() => store.state.model.odinModelResponse.metadata.parameters);
-
-        const reset = () => {
-            const defaultParams: OdinUserType = {};
-            odinParameters.value.forEach((param: OdinParameter) => {
-                if (param.default) {
-                    defaultParams[param.name] = param.default;
-                }
-            });
-            store.commit(`run/${RunMutation.SetParameterValues}`, defaultParams);
-        };
 
         return {
-            reset,
             isFit,
             sensitivityOpen,
             fitTabIsOpen
